Fixed preview for all tabs
