+++ Resources/public/js/components/bulk-price/bulk-price.html
@@ -28,13 +28,13 @@
                     <% _.each(prices, function(price){ %>
                     <tr id="<%= !!price.id ? idPrefix + price.id : 'new' %>" class="row" data-id="<%= !!price.id ? price.id : '' %>">
                         <td>
-                            <input class="form-element husky-validate align-right minimumQuantity preview-update" type="text" value="<%= price.minimumQuantity %>"/>
+                            <input class="form-element husky-validate align-right minimumQuantity" type="text" value="<%= price.minimumQuantity %>"/>
                         </td>
                         <td>
                             <span class="grey-font"><%= translate('product.bulkprice.andMore') %></span>
                         </td>
                         <td>
-                            <input class="form-element husky-validate align-right price preview-update" type="text" value="<%= price.price %>"/>
+                            <input class="form-element husky-validate align-right price" type="text" value="<%= price.price %>"/>
                         </td>
                         <td class="grey-font"><%= currency.code %></td>
                     </tr>
