+++ Resources/public/js/components/products/components/edit/main.js
@@ -7,11 +7,26 @@
  * with this source code in the file LICENSE.
  */
 
-define(['config', 'suluproduct/util/locale-util'], function(Config, LocaleUtil) {
+define([
+    'config',
+    'suluproduct/util/locale-util',
+    'services/sulupreview/preview'
+], function(Config, LocaleUtil, Preview) {
 
     'use strict';
 
     return {
+        layout: function() {
+            return {
+                content: {
+                    width: 'fixed',
+                    leftSpace: true,
+                    rightSpace: true
+                },
+                sidebar: (!!this.options.id) ? 'max' : false
+            }
+        },
+
         header: function() {
             return {
                 toolbar: {
@@ -55,6 +70,22 @@ define(['config', 'suluproduct/util/locale-util'], function(Config, LocaleUtil) 
                     url: '/admin/content-navigations?alias=' + this.options.productType
                 }
             };
+        },
+
+        initialize: function() {
+            if (!this.options.id) {
+                return;
+            }
+
+            this.preview = Preview.initialize({});
+            this.preview.start(
+                'Sulu\\Bundle\\ProductBundle\\Entity\\ProductTranslation',
+                this.options.id,
+                this.options.locale,
+                this.options.data
+            );
+
+            this.preview.bindDomEvents(this.$el);
         }
     };
 });
