+++ Resources/public/js/components/products/components/edit/media/main.js
@@ -228,7 +228,6 @@ define([
         };
 
     return {
-        view: true,
         templates: ['/admin/product/template/product/documents'],
 
         /**
