+++ Resources/views/Template/product.form.html.twig
@@ -20,8 +20,12 @@
         <div class="grid-col-6">
             <div class="form-group">
                 <label for="number" class="required"><%= translate('product.number') %></label>
-                <input class="form-element" type="text" id="number" data-validation-required="true"
-                       data-validation-max-length="255" data-mapper-property="number"/>
+                <input type="text"
+                       id="number"
+                       class="form-element"
+                       data-mapper-property="number"
+                       data-validation-required="true"
+                       data-validation-max-length="255"/>
             </div>
         </div>
         <div class="grid-col-6">
@@ -106,15 +110,21 @@
         <div class="grid-col-6">
             <div class="form-group">
                 <label for="deliveryTime"><%= translate('product.deliveryTime') %></label>
-                <input class="form-element" type="text" id="deliveryTime"
-                       data-mapper-property="deliveryTime" data-type="decimal"/>
+                <input class="form-element"
+                       type="text"
+                       id="deliveryTime"
+                       data-mapper-property="deliveryTime"
+                       data-type="decimal"/>
             </div>
         </div>
         <div class="grid-col-6">
             <div class="form-group">
                 <label for="minimumOrderQuantity"><%= translate('product.minimumOrderQuantity') %></label>
-                <input class="form-element" type="text" id="minimumOrderQuantity"
-                       data-mapper-property="minimumOrderQuantity" data-type="decimal"/>
+                <input class="form-element"
+                       type="text"
+                       id="minimumOrderQuantity"
+                       data-mapper-property="minimumOrderQuantity"
+                       data-type="decimal"/>
             </div>
         </div>
     </div>
@@ -123,14 +133,18 @@
         <div class="grid-col-6">
             <div class="form-group">
                 <label for="manufacturer"><%= translate('products.manufacturer') %></label>
-                <input class="form-element" type="text" id="manufacturer"
+                <input class="form-element"
+                       type="text"
+                       id="manufacturer"
                        data-mapper-property="manufacturer"/>
             </div>
         </div>
         <div class="grid-col-6">
             <div class="form-group">
                 <label for="globalTradeItemNumber"><%= translate('product.globalTradeItemNumber') %></label>
-                <input class="form-element" type="text" id="globalTradeItemNumber"
+                <input class="form-element"
+                       type="text"
+                       id="globalTradeItemNumber"
                        data-mapper-property="globalTradeItemNumber"/>
             </div>
         </div>
